restructure monomorphization pass
