+++ src/passes/monomorphize/mod.rs
@@ -1,0 +1,5 @@
+mod mono;
+mod rewriter;
+
+pub use mono::Monomorphize;
+pub(super) use rewriter::Rewriter;
